backend/requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
matplotlib
plotly

requests
python-binance
pybit>=5.9.0

jupyter

python-dotenv

websocket-client
flask>=2.2
flask-socketio>=5.3
